# docs/openapi.yaml
openapi: 3.0.3
info:
  title: StealADeal Backend API
  version: 0.1.0
  description: |
    OpenAPI contract for the current H2-backed development backend.
    This spec reflects the implemented dealer, vehicle, lead, appointment,
    and dashboard endpoints.
servers:
  - url: http://localhost:8282
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/dealers:
    get:
      tags: [Dealers]
      summary: List dealers
      operationId: listDealers
      security: []
      responses:
        '200':
          description: Dealers returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Dealer'
    post:
      tags: [Dealers]
      summary: Create dealer
      operationId: createDealer
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDealerRequest'
      responses:
        '201':
          description: Dealer created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dealer'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/dealers/{dealerId}:
    get:
      tags: [Dealers]
      summary: Get dealer by id
      operationId: getDealer
      security: []
      parameters:
        - $ref: '#/components/parameters/DealerId'
      responses:
        '200':
          description: Dealer returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dealer'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Dealers]
      summary: Update dealer details
      operationId: updateDealer
      parameters:
        - $ref: '#/components/parameters/DealerId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDealerRequest'
      responses:
        '200':
          description: Dealer updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dealer'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/deals:
    get:
      tags: [Deals]
      summary: List deals
      operationId: listDeals
      parameters:
        - in: query
          name: vehicleId
          schema:
            type: integer
            format: int64
        - in: query
          name: stage
          schema:
            $ref: '#/components/schemas/DealStage'
      responses:
        '200':
          description: Deals returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Deal'
    post:
      tags: [Deals]
      summary: Create a new deal
      operationId: createDeal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDealRequest'
      responses:
        '201':
          description: Deal created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deal'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/deals/{dealId}:
    get:
      tags: [Deals]
      summary: Get deal by id
      operationId: getDeal
      parameters:
        - $ref: '#/components/parameters/DealId'
      responses:
        '200':
          description: Deal returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deal'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/inbox/dealers/{dealerId}:
    get:
      tags: [Inbox]
      summary: Get dealer inbox
      operationId: getDealerInbox
      parameters:
        - $ref: '#/components/parameters/DealerId'
      responses:
        '200':
          description: Dealer inbox returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DealerInbox'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/dealers/{dealerId}/portal:
    get:
      tags: [Portal]
      summary: Get dealer portal overview and pipeline snapshot
      operationId: getDealerPortal
      parameters:
        - $ref: '#/components/parameters/DealerId'
      responses:
        '200':
          description: Dealer portal returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DealerPortal'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/dealers/{dealerId}/portal/deals:
    get:
      tags: [Portal]
      summary: List dealer portal deals
      operationId: listDealerPortalDeals
      parameters:
        - $ref: '#/components/parameters/DealerId'
        - in: query
          name: stage
          schema:
            $ref: '#/components/schemas/DealStage'
      responses:
        '200':
          description: Dealer portal deals returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PortalDealItem'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/dealers/{dealerId}/portal/leads:
    get:
      tags: [Portal]
      summary: List dealer portal leads
      operationId: listDealerPortalLeads
      parameters:
        - $ref: '#/components/parameters/DealerId'
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/LeadStatus'
      responses:
        '200':
          description: Dealer portal leads returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PortalLeadItem'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/dealers/{dealerId}/portal/appointments:
    get:
      tags: [Portal]
      summary: List dealer portal appointments
      operationId: listDealerPortalAppointments
      parameters:
        - $ref: '#/components/parameters/DealerId'
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/AppointmentStatus'
      responses:
        '200':
          description: Dealer portal appointments returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PortalAppointmentItem'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/dealers/{dealerId}/portal/documents:
    get:
      tags: [Portal]
      summary: List dealer portal documents
      operationId: listDealerPortalDocuments
      parameters:
        - $ref: '#/components/parameters/DealerId'
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/DocumentStatus'
      responses:
        '200':
          description: Dealer portal documents returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PortalDocumentItem'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/dealers/{dealerId}/portal/subscription:
    get:
      tags: [Portal]
      summary: Get dealer portal subscription
      operationId: getDealerPortalSubscription
      parameters:
        - $ref: '#/components/parameters/DealerId'
      responses:
        '200':
          description: Dealer portal subscription returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortalSubscription'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Portal]
      summary: Update dealer portal subscription
      operationId: updateDealerPortalSubscription
      parameters:
        - $ref: '#/components/parameters/DealerId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSubscriptionRequest'
      responses:
        '200':
          description: Dealer portal subscription updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortalSubscription'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/dealers/{dealerId}/portal/invoices:
    get:
      tags: [Portal]
      summary: List dealer portal invoices
      operationId: listDealerPortalInvoices
      parameters:
        - $ref: '#/components/parameters/DealerId'
      responses:
        '200':
          description: Dealer portal invoices returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PortalInvoice'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-token
  parameters:
    DealerId:
      in: path
      name: dealerId
      required: true
      schema:
        type: integer
        format: int64
...
    DealId:
      in: path
      name: dealId
      required: true
      schema:
        type: integer
        format: int64
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    Dealer:
      type: object
      required: [id, name, licenseNumber, city, state, approved]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        licenseNumber:
          type: string
        city:
          type: string
        state:
          type: string
        approved:
          type: boolean
    CreateDealerRequest:
      type: object
      required: [name, licenseNumber, city, state]
      properties:
        name:
          type: string
        licenseNumber:
          type: string
        city:
          type: string
        state:
          type: string
          pattern: '^[A-Za-z]{2}$'
    UpdateDealerRequest:
      allOf:
        - $ref: '#/components/schemas/CreateDealerRequest'
...
    Deal:
      type: object
      required:
        - id
        - vehicleId
        - vehicleTitle
        - buyerName
        - buyerEmail
        - buyerPhone
        - buyerAddressLine1
        - buyerCity
        - buyerState
        - buyerPostalCode
        - fulfillmentType
        - tradeIn
        - tradeInOffer
        - vehiclePrice
        - taxAmount
        - registrationFee
        - documentationFee
        - deliveryFee
        - discountAmount
        - depositAmount
        - depositPaid
        - totalAmount
        - stage
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          format: int64
        vehicleId:
          type: integer
          format: int64
        vehicleTitle:
          type: string
        buyerName:
          type: string
        buyerEmail:
          type: string
          format: email
        buyerPhone:
          type: string
        buyerAddressLine1:
          type: string
        buyerAddressLine2:
          type: string
          nullable: true
        buyerCity:
          type: string
        buyerState:
          type: string
        buyerPostalCode:
          type: string
        fulfillmentType:
          $ref: '#/components/schemas/FulfillmentType'
        fulfillmentStatus:
          $ref: '#/components/schemas/FulfillmentStatus'
        fulfillmentScheduledAt:
          type: string
          format: date-time
          nullable: true
        fulfillmentLocation:
          type: string
          nullable: true
        fulfillmentNotes:
          type: string
          nullable: true
        tradeIn:
          type: boolean
        tradeInVin:
          type: string
          nullable: true
        tradeInMileage:
          type: integer
          nullable: true
        tradeInOffer:
          type: number
          format: double
        vehiclePrice:
          type: number
          format: double
        taxAmount:
          type: number
          format: double
        registrationFee:
          type: number
          format: double
        documentationFee:
          type: number
          format: double
        deliveryFee:
          type: number
          format: double
        discountAmount:
          type: number
          format: double
        depositAmount:
          type: number
          format: double
        depositPaid:
          type: boolean
        totalAmount:
          type: number
          format: double
        stage:
          $ref: '#/components/schemas/DealStage'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateDealRequest:
      type: object
      required:
        - vehicleId
        - buyerName
        - buyerEmail
        - buyerPhone
        - buyerAddressLine1
        - buyerCity
        - buyerState
        - buyerPostalCode
        - fulfillmentType
        - tradeIn
        - tradeInOffer
        - deliveryFee
        - discountAmount
      properties:
        vehicleId:
          type: integer
          format: int64
        buyerName:
          type: string
        buyerEmail:
          type: string
          format: email
        buyerPhone:
          type: string
        buyerAddressLine1:
          type: string
        buyerAddressLine2:
          type: string
          nullable: true
        buyerCity:
          type: string
        buyerState:
          type: string
          pattern: '^[A-Za-z]{2}$'
        buyerPostalCode:
          type: string
          pattern: '^[0-9]{5}(-[0-9]{4})?$'
        fulfillmentType:
          $ref: '#/components/schemas/FulfillmentType'
        tradeIn:
          type: boolean
        tradeInVin:
          type: string
          pattern: '^[A-HJ-NPR-Z0-9]{17}$'
          nullable: true
        tradeInMileage:
          type: integer
          minimum: 0
          nullable: true
        tradeInOffer:
          type: number
          format: double
          minimum: 0
        deliveryFee:
          type: number
          format: double
          minimum: 0
        discountAmount:
          type: number
          format: double
          minimum: 0
...
    PortalDealItem:
      type: object
      required: [dealId, vehicleId, vehicleTitle, buyerName, stage, depositPaid, readyForHandoff, blockers, nextAction, updatedAt]
      properties:
        dealId:
          type: integer
          format: int64
        vehicleId:
          type: integer
          format: int64
        vehicleTitle:
          type: string
        buyerName:
          type: string
        stage:
          $ref: '#/components/schemas/DealStage'
        depositPaid:
          type: boolean
        readyForHandoff:
          type: boolean
        blockers:
          type: array
          items:
            type: string
        nextAction:
          type: string
        updatedAt:
          type: string
          format: date-time
    PortalLeadItem:
      type: object
      required: [id, vehicleId, vehicleTitle, buyerName, buyerEmail, status, message, createdAt]
      properties:
        id:
          type: integer
          format: int64
        vehicleId:
          type: integer
          format: int64
        vehicleTitle:
          type: string
        buyerName:
          type: string
        buyerEmail:
          type: string
          format: email
        status:
          $ref: '#/components/schemas/LeadStatus'
        message:
          type: string
        createdAt:
          type: string
          format: date-time
    PortalAppointmentItem:
      type: object
      required: [id, vehicleId, vehicleTitle, buyerName, buyerEmail, type, status, scheduledAt, createdAt]
      properties:
        id:
          type: integer
          format: int64
        vehicleId:
          type: integer
          format: int64
        vehicleTitle:
          type: string
        buyerName:
          type: string
        buyerEmail:
          type: string
          format: email
        type:
          $ref: '#/components/schemas/AppointmentType'
        status:
          $ref: '#/components/schemas/AppointmentStatus'
        scheduledAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    PortalDocumentItem:
      type: object
      required: [id, dealId, vehicleTitle, buyerName, type, status, fileName, updatedAt]
      properties:
        id:
          type: integer
          format: int64
        dealId:
          type: integer
          format: int64
        vehicleTitle:
          type: string
        buyerName:
          type: string
        type:
          $ref: '#/components/schemas/DocumentType'
        status:
          $ref: '#/components/schemas/DocumentStatus'
        fileName:
          type: string
        updatedAt:
          type: string
          format: date-time
    PortalSubscription:
      type: object
      required: [id, dealerId, plan, status, monthlyPrice, currentPeriodStart, currentPeriodEnd, autoRenew, updatedAt]
      properties:
        id:
          type: integer
          format: int64
        dealerId:
          type: integer
          format: int64
        plan:
          $ref: '#/components/schemas/SubscriptionPlan'
        status:
          $ref: '#/components/schemas/SubscriptionStatus'
        monthlyPrice:
          type: number
          format: double
        currentPeriodStart:
          type: string
          format: date-time
        currentPeriodEnd:
          type: string
          format: date-time
        autoRenew:
          type: boolean
        updatedAt:
          type: string
          format: date-time
    UpdateSubscriptionRequest:
      type: object
      required: [plan, status, autoRenew]
      properties:
        plan:
          $ref: '#/components/schemas/SubscriptionPlan'
        status:
          $ref: '#/components/schemas/SubscriptionStatus'
        autoRenew:
          type: boolean
    PortalInvoice:
      type: object
      required: [id, invoiceNumber, status, amount, periodStart, periodEnd, dueAt, createdAt]
      properties:
        id:
          type: integer
          format: int64
        invoiceNumber:
          type: string
        status:
          $ref: '#/components/schemas/InvoiceStatus'
        amount:
          type: number
          format: double
        periodStart:
          type: string
          format: date-time
        periodEnd:
          type: string
          format: date-time
        dueAt:
          type: string
          format: date-time
        paidAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
...
    DealerPortal:
      type: object
      required: [overview, pipeline, queueSummary, recentDeals, recentActivity]
      properties:
        overview:
          $ref: '#/components/schemas/PortalOverview'
        pipeline:
          $ref: '#/components/schemas/PipelineSnapshot'
        queueSummary:
          $ref: '#/components/schemas/QueueSummary'
        recentDeals:
          type: array
          items:
            $ref: '#/components/schemas/PortalDealItem'
        recentActivity:
          type: array
          items:
            $ref: '#/components/schemas/PortalActivityItem'
...
    DealerInbox:
      type: object
      required: [inbox, queueSummary]
      properties:
        inbox:
          $ref: '#/components/schemas/ParticipantInbox'
        queueSummary:
          $ref: '#/components/schemas/QueueSummary'
...
    LeadStatus:
      type: string
      enum:
        - NEW
        - CONTACTED
        - QUALIFIED
        - CLOSED
...
    AppointmentStatus:
      type: string
      enum:
        - REQUESTED
        - CONFIRMED
        - COMPLETED
        - CANCELED
...
    DealStage:
      type: string
      enum:
        - INITIATED
        - OFFER_SENT
        - BUYER_CONFIRMED
        - DEPOSIT_PAID
        - DOCUMENTS_PENDING
        - READY_FOR_HANDOFF
        - COMPLETED
        - CANCELED
...
    DocumentStatus:
      type: string
      enum:
        - REQUESTED
        - UPLOADED
        - APPROVED
        - REJECTED
...